requirements.txt
python-dotenv
ipython
requests
pytest
